# Make RSpec/SubjectStub stricter