@pirj 2022-01-16
Some cases are no longer detected, probably due to mistaken optimizations:
```ruby
  RSpec.describe Spree.user_class, type: :model do
    context "reporting" do
      def load_orders
        allow(subject).to receive(:spree_orders).and_return(double(complete: orders))
      end
    end
  end
```

I'm checking.